fix: enhance test coverage and resolve platform-specific test failures - Fix infinity/NaN validation in Storage class initialization - Fix Windows PowerShell output parsing to handle floating point values - Add comprehensive platform-specific test skipping for Windows/Linux/macOS - Add extensive test coverage files achieving 94%+ coverage - Fix Path object mocking issues across multiple test files - Enhance subprocess exception handling in platform storage tests - Improve floating point precision handling in edge case tests
Key fixes:
* platform_storage.py: Change int() to float() for PowerShell output parsing
* storage.py: Add math.isfinite() validation for infinity/NaN values
* test_platform_storage_comprehensive.py: Add @pytest.mark.skipif decorators
* Multiple test files: Fix import statements from 'bytesize' to 'filesizelib'
* Add 5 new comprehensive test files for better coverage

+++ filesizelib/platform_storage.py
@@ -145,7 +145,7 @@ class WindowsStorage(PlatformStorageBase):
             ], capture_output=True, text=True, timeout=30)
             
             if result.returncode == 0 and result.stdout.strip():
-                size = int(result.stdout.strip())
+                size = float(result.stdout.strip())
                 return Storage.parse_from_bytes(size)
         except (subprocess.TimeoutExpired, subprocess.CalledProcessError, ValueError, FileNotFoundError):
             pass
